UPSTREAM: test: Guarantee platform cleanup in flash_and_run_test
Implement context manager methods (__enter__ and __exit__) on the
Platform base class and wrap test execution in flash_and_run_test() with
'with platform:' so platform.cleanup() is unconditionally called even if
interrupted by Ctrl+C or an exception.

Also guard self.process.kill() with 'if self.process:' and call
self.process.wait() in Renode.cleanup() so cleanup reaps the child
process and can safely run before a process is spawned.

Original-BUG=b:368684364
Original-TEST=./test/run_device_tests.py --renode --board=bloonchipper --tests production_app_test
Original-TAG=agy
Original-CONV=cf1f9be3-70b1-41f9-927b-39856a5e7ab1
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/8419582
Original-Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Original-Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Original-Tested-by: Jeremy Bettis <jbettis@chromium.org>
Original-Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
GitOrigin-RevId: 5c1e267b6449aefc7c77e0653c0f8467d46fb6b9
Cr-Build-Id: 8670434206165718241
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8670434206165718241
Copybot-Job-Name: fingerprint-ec-legacy-copybot-downstream
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/8420064
Commit-Queue: Tom Hughes <tomhughes@chromium.org>
Tested-by: Tom Hughes <tomhughes@chromium.org>

+++ test/run_device_tests.py
@@ -314,6 +314,12 @@ class Platform(ABC):
     def cleanup(self) -> None:
         """Clean up after a test run."""
 
+    def __enter__(self) -> Platform:
+        return self
+
+    def __exit__(self, exc_type, exc_value, traceback) -> None:
+        self.cleanup()
+
     @abstractmethod
     def skip_test(
         self, test_config: TestConfig, board_config: BoardConfig, zephyr: bool
@@ -462,7 +468,10 @@ class Renode(Platform):
         return True
 
     def cleanup(self) -> None:
-        self.process.kill()
+        if self.process:
+            self.process.kill()
+            self.process.wait()
+            self.process = None
 
     def _skip_test_bloonchipper(
         self, test_config: TestConfig, zephyr: bool
@@ -1944,60 +1953,57 @@ def flash_and_run_test(
         logging.info("Build complete. Skipping execution due to --build-only.")
         return True
 
-    # Get the console file before flashing to listen ASAP after flashing.
-    console_pty = platform.get_console(board_config)
+    with platform:
+        # Get the console file before flashing to listen ASAP after flashing.
+        console_pty = platform.get_console(board_config)
 
-    # flash test binary
-    if not platform.flash(
-        board_config,
-        image_path,
-        args.flasher,
-        args.remote,
-        args.jlink_port,
-        test.test_name,
-        test.enable_hw_write_protect,
-        args.zephyr,
-    ):
-        logging.debug("Flashing failed")
-        return False
-
-    with ExitStack() as stack:
-        if args.remote and args.console_port:
-            console_socket = stack.enter_context(
-                socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            )
-            console_socket.connect((args.remote, args.console_port))
-            console_socket.setblocking(False)
-            console = stack.enter_context(
-                console_socket.makefile(mode="rwb", buffering=0)
-            )
-        else:
-            # pylint: disable-next=consider-using-with
-            console_file = open(console_pty, "wb+", buffering=0)
-            console = stack.enter_context(console_file)
-            os.set_blocking(console.fileno(), False)
-
-        platform.hw_write_protect(test.enable_hw_write_protect)
-
-        if test.toggle_power:
-            power_cycle(platform, board_config)
-        else:
-            # In some cases flash_ec leaves the board off, so just ensure it is on
-            platform.power(board_config, power_on=True)
-
-        # run the test
-        logging.info('Running test: "%s"', test.config_name)
-
-        ret = run_test(
-            test,
+        # flash test binary
+        if not platform.flash(
             board_config,
-            console,
-            zephyr=args.zephyr,
-        )
+            image_path,
+            args.flasher,
+            args.remote,
+            args.jlink_port,
+            test.test_name,
+            test.enable_hw_write_protect,
+            args.zephyr,
+        ):
+            logging.debug("Flashing failed")
+            return False
 
-        platform.cleanup()
+        with ExitStack() as stack:
+            if args.remote and args.console_port:
+                console_socket = stack.enter_context(
+                    socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+                )
+                console_socket.connect((args.remote, args.console_port))
+                console_socket.setblocking(False)
+                console = stack.enter_context(
+                    console_socket.makefile(mode="rwb", buffering=0)
+                )
+            else:
+                # pylint: disable-next=consider-using-with
+                console_file = open(console_pty, "wb+", buffering=0)
+                console = stack.enter_context(console_file)
+                os.set_blocking(console.fileno(), False)
 
-        return ret
+            platform.hw_write_protect(test.enable_hw_write_protect)
+
+            if test.toggle_power:
+                power_cycle(platform, board_config)
+            else:
+                # In some cases flash_ec leaves the board off, so just ensure it is on
+                platform.power(board_config, power_on=True)
+
+            # run the test
+            logging.info('Running test: "%s"', test.config_name)
+
+            return run_test(
+                test,
+                board_config,
+                console,
+                zephyr=args.zephyr,
+            )
 
 
 def write_json_results(
